Implemented some asBC_CALL behavior, modify script virtual function behavior
Before this, the jitFunction of the script function would be set to the jitentry (which is meant to be the entry point for asBC_JitEntry) OR to the generated function directly (which is NOT meant to be used as a JitEntry entry point) if it was a method.

This was not a robust way to do this, however, since constructors can be used as an entry point, and the user _might_ (or not, unsure) be able to call into a method from the application directly.

Now the jitFunction is always set to be the actual jitentry and the function itself is defined as userdata to the script under a category hardcoded in modulecommon.hpp. While this is not ideal, partly for performance reasons, this is certainly much more reliable.