Use the script parameter types as the function type
Instead of taking a single %params for every translated function, we take in real parameters of the matching types.

There are several main reasons:
- Debugging is likely less affected
- Codegen should be better in general: First, LLVM can probably do better assumptions on how the function params are used; second, this should allow less memory accesses.
- Factory functions behave quite differently from most functions and partially use their parameters to forward to the call as an optimization. This would not be easy to fix without this.

We can also now remove the old optimization thunk thing as it was only relevant when transforming the function signature made any sense

This is a first commit which does only support trivial functions.